자료실 crud 전체 기능 코드 수정 완료, 개인업무일지 작성 기능 작업 진행중

+++ src/main/java/kr/co/bizcore/v1/svc/BoardService.java
@@ -197,7 +197,7 @@ public class BoardService extends Svc{
             if(fileList != null && fileList.size() > 0) for(x = 0 ; x < fileList.size() ; x++){
                 savedName = fileList.get(x).getSavedName();
                 for(Object obj : removeFiles){
-                    if(savedName.equals(obj)){
+                    if(savedName != null && savedName.equals(obj)){
                         targetFile = new File(path + s + fileList.get(x).getSavedName());
                         if(targetFile.exists()) targetFile.delete();
                         boardMapper.deleteFileboxAttachedFile(article.getNo() + "", compId, savedName);
